+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # Version 1.0.5
 
 * fix `_const` for libvips 8.5 [John Cupitt]
+* add `scaleimage`, the scale operation renamed to avoid a clash with the
+  `scale` property [John Cupitt]
 
 # Version 1.0.4